reboot server during provision when required
